refactor: utils folder with waitForElement and toggleRecordingFromContentScript
