+++ apps/extension/contents/poe.tsx
@@ -1,13 +1,9 @@
-import { sendToBackground } from '@plasmohq/messaging';
 import { useStorage } from '@plasmohq/storage/hook';
-import { WhisperingError, recorderStateToIcons, type RecorderState } from '@repo/shared';
+import { recorderStateToIcons, type RecorderState } from '@repo/shared';
 import cssText from 'data-text:~/style.css';
-import { Effect } from 'effect';
 import type { PlasmoCSConfig, PlasmoGetInlineAnchor, PlasmoGetStyle } from 'plasmo';
-import { renderErrorAsNotification } from '~lib/errors';
-import { NotificationServiceContentLive } from '~lib/services/NotificationServiceContentLive';
 import { STORAGE_KEYS } from '~lib/services/extension-storage';
-import type * as ToggleRecording from '../background/messages/contents/toggleRecording';
+import { toggleRecordingFromContentScript } from './utils/toggleRecordingFromContentScript';
 
 export const getInlineAnchor: PlasmoGetInlineAnchor = async () => {
 	const selector = 'div.GrowingTextArea_growWrap__im5W3';
@@ -34,27 +30,6 @@ export const getStyle: PlasmoGetStyle = () => {
 	style.textContent = cssText.replaceAll(':root', ':host(plasmo-csui)');
 	return style;
 };
-
-const toggleRecordingFromContentScript = () =>
-	Effect.tryPromise({
-		try: () =>
-			sendToBackground<ToggleRecording.RequestBody, ToggleRecording.ResponseBody>({
-				name: 'contents/toggleRecording',
-			}),
-		catch: (error) =>
-			new WhisperingError({
-				title: `Unable to toggle recording via background service worker`,
-				description:
-					error instanceof Error
-						? error.message
-						: 'There was likely an issue sending the message to the background service worker from the contentscript.',
-				error,
-			}),
-	}).pipe(
-		Effect.catchAll(renderErrorAsNotification),
-		Effect.provide(NotificationServiceContentLive),
-		Effect.runPromise,
-	);
 
 function RecorderStateAsIcon() {
 	const [recorderState] = useStorage<RecorderState>(STORAGE_KEYS.RECORDER_STATE, 'IDLE');
