+++ apps/extension/contents/utils/waitForElement.ts
